+++ packages/stryker/test/helpers/producers.ts
@@ -10,6 +10,8 @@ import TestableMutant from '../../src/TestableMutant';
 import SourceFile from '../../src/SourceFile';
 import TranspiledMutant from '../../src/TranspiledMutant';
 import { Logger } from 'log4js';
+import { FileCoverageData } from 'istanbul-lib-coverage';
+import { CoverageMaps } from '../../src/transpiler/CoverageInstrumenterTranspiler';
 
 export type Mock<T> = {
   [P in keyof T]: sinon.SinonStub;
@@ -110,6 +112,20 @@ export const fileLocation = factory<FileLocation>({
   start: { line: 0, column: 0 }, end: { line: 0, column: 0 }
 });
 
+export const coverageMaps = factoryMethod<CoverageMaps>(() => ({
+  statementMap: {},
+  fnMap: {}
+}));
+
+export const fileCoverageData = factoryMethod<FileCoverageData>(() => ({
+  path: '',
+  statementMap: {},
+  b: {},
+  branchMap: {},
+  f: {},
+  fnMap: {},
+  s: {}
+}));
 
 export const testFramework = factory<TestFramework>({
   beforeEach(codeFragment: string) { return `beforeEach(){ ${codeFragment}}`; },