+++ packages/stryker-api/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.12.0"></a>
+# 0.12.0 (2017-12-21)
+
+
+### Features
+
+* **cvg analysis:** New coverage instrumenter ([#550](https://github.com/stryker-mutator/stryker/issues/550)) ([2bef577](https://github.com/stryker-mutator/stryker/commit/2bef577))
+* **typescript:** Add version check ([#449](https://github.com/stryker-mutator/stryker/issues/449)) ([a780189](https://github.com/stryker-mutator/stryker/commit/a780189)), closes [#437](https://github.com/stryker-mutator/stryker/issues/437)
+
+
+
+
 <a name="0.11.0"></a>
 # [0.11.0](https://github.com/stryker-mutator/stryker/compare/stryker-api@0.10.0...stryker-api@0.11.0) (2017-10-24)